% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHomeRangeData.R
\name{GetHomeRangeData}
\alias{GetHomeRangeData}
\title{GetHomeRangeData}
\usage{
GetHomeRangeData(IncludeReferences = FALSE)
}
\arguments{
\item{IncludeReferences}{option to include the source references directly to the HomeRange data, can be set to TRUE or FALSE (default: FALSE)}
}
\value{
The HomeRange dataset
}
\description{
Downloads the latest HomeRange dataset and loads it as a data.frame into R.
}
\examples{
library(HomeRange)
HomeRangeData <- GetHomeRangeData()
}
