% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomeRangeVersion.R
\name{HomeRangeVersion}
\alias{HomeRangeVersion}
\title{HomeRangeVersion}
\usage{
HomeRangeVersion()
}
\description{
This function lists the version of the package and database.
}
\examples{
library(HomeRange)

# view HomeRange metadata
HomeRangeVersion()
}
