% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeStatTable.R
\name{MakeStatTable}
\alias{MakeStatTable}
\title{MakeStatTable}
\usage{
MakeStatTable(HomeRangeData, ReturnTable = FALSE)
}
\arguments{
\item{HomeRangeData}{the HomeRange dataset (can be retrieved by running: GetHomeRangeData())}

\item{ReturnTable}{option to return the table with statistics as a data.frame, TRUE or FALSE (default: FALSE)}
}
\description{
Creates a table for the HomeRange dataset using the gridExtra package.
}
\examples{
library(HomeRange)
HomeRangeData <- GetHomeRangeData()
MakeStatTable(HomeRangeData)
}
