% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHistogram.R
\name{PlotHistogram}
\alias{PlotHistogram}
\title{PlotHistogram}
\usage{
PlotHistogram(HomeRangeData, Labelvjust = 2, PanelLabelSize = 4)
}
\arguments{
\item{HomeRangeData}{the HomeRange dataset (can be retrieved by running: GetHomeRangeData())}

\item{Labelvjust}{value to adjust the labeling of the histogram panels, default is 2}

\item{PanelLabelSize}{value to adjust the size of the panel labeling, default is 4}
}
\description{
Plot HomeRange data specially on a global map.
}
\examples{
library(HomeRange)
HomeRangeData <- GetHomeRangeData()
PlotHistogram(HomeRangeData)
}
