% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMap.R
\name{PlotMap}
\alias{PlotMap}
\title{PlotMap}
\usage{
PlotMap(HomeRangeData, DotColor = "red")
}
\arguments{
\item{HomeRangeData}{the HomeRange dataset (can be retrieved by running: GetHomeRangeData())}

\item{DotColor}{string to set to color of the home-range locations, default is "red"}
}
\description{
Plot HomeRange data specially on a global map.
}
\examples{
library(HomeRange)
HomeRangeData <- GetHomeRangeData()
PlotMap(HomeRangeData)
}
