% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewMetaData.R
\name{ViewMetaData}
\alias{ViewMetaData}
\title{ViewMetaData}
\usage{
ViewMetaData()
}
\description{
This function opens the metadata pdf file in the browser.
}
\examples{
library(HomeRange)

# view HomeRange metadata
ViewMetaData()
}
